Use long to avoid data loss on 64-bit

+++ tankbobs/src/common/m_world.cpp
@@ -1150,7 +1150,7 @@ int w_getIndex(lua_State *L)
 	b2Body *body = reinterpret_cast<b2Body *>(lua_touserdata(L, 1));
 	lua_settop(L, 0);
 
-	lua_pushinteger(L, reinterpret_cast<int>(body->GetUserData()));
+	lua_pushinteger(L, reinterpret_cast<long>(body->GetUserData()));
 
 	return 1;
 }
